requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
boto3>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
opencv-python>=4.10.0
httpx>=0.27.0
numpy>=2.0.0
openai-whisper>=20240930
asyncpg>=0.30.0
